requirements.txt
pip>=9.0.1
Flask==0.12.2
requests
requests_cache
plotly
cassandra-driver
Flask-HTTPAuth
datetime
passlib
flask-jwt-extended
